+++ skills/datarobot-agent-assist/agent-assist-simulate/scripts/native_scenarios.py
@@ -25,7 +25,7 @@ from artifacts import (
     write_criteria,
     write_json,
 )
-from contracts import (
+from swarm_contracts import (
     ConvergenceConfig,
     EvaluationConfig,
     ExecutionConfig,
