+++ skills/datarobot-agent-assist/agent-assist-simulate/scripts/native_swarm.py
@@ -28,7 +28,7 @@ from artifacts import (
     load_spec,
     write_json,
 )
-from contracts import (
+from swarm_contracts import (
     AgentSpec,
     Scenario,
     ScenarioResult,
