+++ skills/datarobot-agent-assist/agent-assist-simulate/scripts/swarm_contracts.py
