+++ tests/integration/test_native_convergence.py
@@ -21,7 +21,7 @@ SCRIPT_DIR = (
 SCRIPT_PATH = SCRIPT_DIR / "native_convergence.py"
 sys.path.insert(0, str(SCRIPT_DIR))
 artifacts = importlib.import_module("artifacts")
-contracts = importlib.import_module("contracts")
+contracts = importlib.import_module("swarm_contracts")
 native = importlib.import_module("native_convergence")
