+++ tests/integration/test_native_swarm.py
@@ -22,7 +22,7 @@ SCRIPT_DIR = (
 SCRIPT_PATH = SCRIPT_DIR / "native_swarm.py"
 sys.path.insert(0, str(SCRIPT_DIR))
 artifacts = importlib.import_module("artifacts")
-contracts = importlib.import_module("contracts")
+contracts = importlib.import_module("swarm_contracts")
 execution = importlib.import_module("native_execution")
 swarm = importlib.import_module("native_swarm")
